Ax passes SMHasher 2!
Speed is good, though not quite as good as SMHasher 3's version. Fast in bulk and on small keys. I had to change ax_mix_stream_bulk() to use different rotations and and add instead of subtract. Other changes are mostly for readability, though the bad seeds test for all seeds does test the same default problematic seeds as the single-seed test, now (it didn't test 48 or 128 before).